% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_azure.R
\name{get_item_azure}
\alias{get_item_azure}
\title{Get an item listing from Azure OneDrive}
\usage{
get_item_azure(drive, path)
}
\arguments{
\item{drive}{A pointer to an ms_drive object from \code{Microsoft365R}}

\item{path}{The path to the file.}
}
\value{
An \code{ms_drive_item} object containing the requested data.
}
\description{
Get an item \code{ms_drive_item} object from Azure given a path.
}
\details{
The cloud operations of Azure split between shared and owned files.
Collaborative projects need code that remains consistent and simple
in retrieval of files. This function deals with the complexity of
find the item between owned and shared files and allows for different
users to utilize the same simple function.
}
\examples{
\dontrun{
library(Microsoft365R)
drive <- get_business_onedrive() # or drive_type="device_code")
item  <- get_item_azure(drive, "/SomeDir/sharedata.csv")
}
}
