% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owned_azure.R
\name{owned_azure}
\alias{owned_azure}
\title{Named list of Microsoft365R items that are owned on OneDrive}
\usage{
owned_azure(drive)
}
\arguments{
\item{drive}{A pointer to an ms_drive object from \code{Microsoft365R}}
}
\value{
Named list of \code{Microsoft365R::ms_drive_item}'s owned by user.
}
\description{
Retrieve a named list of owned OneDrive items given a drive authentication
object.
}
\examples{
\dontrun{
drive <- get_business_onedrive(auth_type="device_code")
names(owned_azure(drive))
}
}
