% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure.R
\name{read_azure}
\alias{read_azure}
\title{Read a data in a file from Azure OneDrive}
\usage{
read_azure(drive, path, FUN = NULL, ...)
}
\arguments{
\item{drive}{A pointer to an ms_drive object from \code{Microsoft365R}}

\item{path}{The path to the file.}

\item{FUN}{The function to read the data into memory. The default will make a guess of supported type from the filename extension.}

\item{\dots}{Additional arguments to pass to the function that reads the data into memory.}
}
\value{
An R object containing the requested data.
}
\description{
This function given a path will read a file that was shared or owned directly
into memory. This is important if the file contains information
that one doesn't want stored to disk, e.g. private health information (PHI)
or private identifiable information (PII).
}
\details{
Unfortunately, if one "mounts" a OneDrive it will by DEFAULT
copy the files locally. The files must be manually turned off to "sync".
}
\examples{
\dontrun{
library(Microsoft365R)
drive <- get_business_onedrive(drive_type="device_code")
data  <- read.shared(drive, "/SomeDir/sharedata.csv")
}
}
