% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_azure.R
\name{shared}
\alias{shared}
\title{Microsoft365R items that are shared on drive}
\usage{
shared(drive)
}
\arguments{
\item{drive}{A pointer to an ms_drive object from \code{Microsoft365R}}
}
\value{
Named list of \code{Microsoft365R::ms_drive_item}'s shared with user.
}
\description{
Retrieve a list of shared OneDrive items given a drive authentication
object.
}
\examples{
\dontrun{
drive <- get_business_onedrive(drive_type="device_code")
names(shared(drive))
}
}
