% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_azure.R
\name{shared_azure}
\alias{shared_azure}
\title{Named list of Microsoft365R items that are shared on OneDrive}
\usage{
shared_azure(drive)
}
\arguments{
\item{drive}{A pointer to an ms_drive object from \code{Microsoft365R}}
}
\value{
Named list of \code{Microsoft365R::ms_drive_item}'s shared with user.
}
\description{
Retrieve a named list of shared OneDrive items given a drive authentication
object.
}
\examples{
\dontrun{
drive <- get_business_onedrive(auth_type="device_code")
names(shared_azure(drive))
}
}
