% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_azure.R
\name{write_azure}
\alias{write_azure}
\title{Write data from R memory to OneDrive}
\usage{
write_azure(drive, x, path, FUN = NULL, ...)
}
\arguments{
\item{drive}{A pointer to an ms_drive object from \code{Microsoft365R}}

\item{x}{The data to be written.}

\item{path}{The OneDrive path to write too with filename.}

\item{FUN}{The function to read the data into memory. The default will make a
guess of supported type from the filename extension. The first argument to
the function must be the data. The second argument is what would normally be
the file (it will be a \code{\link{rawConnection}}). This function
defaults to NULL and when that is the case the routine will make a guess
based on the file extension, e.g. if the extension is ".csv" it will use
R's \code{\link{write.csv}}.}

\item{\dots}{Additional arguments to pass to the function that writes the file.}
}
\value{
An R object containing the requested data.
}
\description{
This function given a path will write a file to OneDrive from memory.
}
\details{
If \code{FUN} is \code{NULL} it will guess based on file extension. It
supports: arff, csv, dbf, dta, rds, rdata, rec, table, yml, and yaml.
}
\examples{
\dontrun{
library(Microsoft365R)
drive <- get_business_onedrive(auth_type="device_code")
data  <- write_azure(drive, data, "/SomeDir/sharedata.csv")
}
}
\seealso{
\code{\link{link_path}}, \code{\link{owned_azure}}, and \code{\link{shared_azure}}
}
